requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.3
pywebview==5.2
pyinstaller==6.11.1
